+++ src/Dash/layout/DashboardLayout.js
@@ -1,7 +1,7 @@
 import React,{useEffect, useState} from 'react'
 import {Redirect} from 'react-router-dom'
 import { AppContent, AppSidebar, AppFooter, AppHeader } from '../dashboard/index'
-import {COLUMNS} from "../assessor/columns";
+import {COLUMNS} from "../views/assessor/columns";
 import {API, Auth, graphqlOperation} from 'aws-amplify'
 import * as queries from '../../graphql/queries';
 import * as mutations from '../../graphql/mutations'
@@ -9,7 +9,7 @@ import {sendEmail} from '../../Home/shared/functions/AwsFuncs'
 import IdleTimer from "./IdleTimer";
 
 
-const DefaultLayout = (props) => {
+const DashboardLayout = (props) => {
   const {signedIn, signOut, userGroup} = props;
   const [userDetails, setUserDetails] = useState({});
   const [hasQuestionnaire, setHasQuestionnaire ] = useState(false);
@@ -50,6 +50,7 @@ const DefaultLayout = (props) => {
       return () => {
         timer.cleanUp();
       };
+      // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
   
     useEffect(() => { 
@@ -68,6 +69,7 @@ const DefaultLayout = (props) => {
         }
       }
       getDashValues();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
 
     },[])
     
@@ -124,7 +126,7 @@ const DefaultLayout = (props) => {
 
     const getAssessReportHandler = async (userId)=>{
 
-      const response = API.graphql({query:queries.listAssessorReports, variables: {filter: {userId: {contains: userId}}}}).then((arpd)=>{
+      await API.graphql({query:queries.listAssessorReports, variables: {filter: {userId: {contains: userId}}}}).then((arpd)=>{
         if(arpd)
       assRepHandler(arpd.data.listAssessorReports.items[0])
       MILHandler(arpd.data.listAssessorReports.items[0])
@@ -171,7 +173,6 @@ const DefaultLayout = (props) => {
 
     const recommendationsHandler = async (id)=>{
       const recommendations = await listProps("Rec", id)
-      // getApproved(recommendations);
       getPending(recommendations);
 
     }
@@ -271,14 +272,6 @@ const addEventHandler=(task)=>{
   }
 
   
-    function getApproved(recommendations){//scans through the a recommendations array, gets approved recs, and moves those recs to a new array
-            const approvedRecs = recommendations.filter((rec)=>{
-              return rec.isApproved
-            });//filters 
-          
-          console.log("These are the apporved recos", approvedRecs);
-          setApproved(approvedRecs);
-          }
     function getPending(recommendations){//
             const pendingRecs = recommendations.filter((rec)=>{
               return rec.isApproved === false;
@@ -293,13 +286,11 @@ const addEventHandler=(task)=>{
       function approve(rec){
         rec.isApproved = true;
         setRecommendations(recommendations.map((recommendation)=>{
-          // console.log(recommendation.isApproved)
           return recommendation.id === rec.id ? {...recommendation, isApproved: true}: recommendation;
         }))
         if(rec){
           addToApprovedHandler(rec)
         }
-        // console.log("rec is ",rec.isApproved)
 
        }
 
@@ -443,4 +434,4 @@ const addEventHandler=(task)=>{
   )
 }
 
-export default DefaultLayout
+export default DashboardLayout
