+++ src/Dash/views/assessor/Report.js
@@ -1,33 +1,30 @@
 import React, { useState, useEffect} from 'react';
 import {Button, Container} from 'react-bootstrap'
 import { MDBDataTableV5 } from 'mdbreact';
-import {useHistory} from 'react-router-dom'
+import {Link, Redirect} from 'react-router-dom'
 import {COLUMNS} from "./columns.js";
-import * as queries from '../../graphql/queries'
+import * as queries from '../../../graphql/queries'
 import API from '@aws-amplify/api';
 
-function ExpertViewCustList(props){
-  // const {userId} = props;
+function Report(){
  const [checkbox1, setCheckbox1] = useState('');
  const [datatable, setDatatable] = useState('');
  const [hasData, setHasData] = useState(false);
- const [disabled, setDisabled] = useState(true);
- const history = useHistory();
  const showLogs2 = (e) => {
-  setDisabled(false);   
-  setCheckbox1(e);
+     console.log('SHOWLOGS:::', e);
+     setCheckbox1(e);
  };
- 
  useEffect(() => {
+   
    listUsers().then(listOfUsers => {
      let users = [];
-     let completed;
+     var completed;
           
           for(let i in  listOfUsers.data.listUsers.items){
-
             if( listOfUsers.data.listUsers.items[i].userType === "Assessee"){
               completed =  listOfUsers.data.listUsers.items[i];
-              users.push(completed)
+              console.log("This is the approved ",  listOfUsers.data.listUsers.items[i])
+              users.push( completed);
               
             }
            }
@@ -36,17 +33,16 @@ function ExpertViewCustList(props){
                  rows: users
              }
              setDatatable(data);
-         }).finally(()=>{
-             setHasData(true);
-         });
+         })
        },[])
 
-       function assess(){
-        console.log("This is the selected user", checkbox1);
-        history.push("/dash/assess" ,{userId: checkbox1.id})
-       }
+       useEffect(()=>{
+          console.log("Hasdata", datatable)
+          if(datatable){
+            setHasData(true);
+        }},[datatable])
 
-  async function listUsers() {
+async function listUsers() {
       try {
            var userslist = await API.graphql({query: queries.listUsers});
           return userslist;
@@ -54,11 +50,10 @@ function ExpertViewCustList(props){
           console.log("Error:>> ", err);
       }
   }
-
   return (<>
-    {hasData  && (<><Container>
-<h4 className="text-center display-4">Customer list</h4>
-<span>Click on the customer to view answers</span>
+    {hasData  ? (<><Container>
+<h4 className="text-center display-4">Reports</h4>
+<span></span>
 </Container>
     <div className="py-3">
       <MDBDataTableV5
@@ -69,13 +64,16 @@ function ExpertViewCustList(props){
         headCheckboxID='id2'
         bodyCheckboxID='checkboxes2'
         getValueCheckBox={(e) => {
-          showLogs2(e)
-          ;}}
+          showLogs2(e);}}
           />
     </div>
-    <Button className= "ml-5 mb-5" disabled={disabled} onClick={assess} type="submit">Assess</Button>  
-    </>)}
+    <Link to= {{pathname: "/dash/assess" ,state: checkbox1}}><Button className= "ml-5 mb-5"  type="submit">New Report
+  </Button></Link>
+   <div>
+      {checkbox1 && <p>{JSON.stringify(delete checkbox1.checkbox && checkbox1.first_name)}</p>}
+    </div>
+    </>):<Redirect to="/500"/>}
     </>
   );}
  
-  export default ExpertViewCustList; 
+  export default Report;
