+++ src/Layouts.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {  Redirect, Route, Switch } from 'react-router-dom';
 
 //Containers
-import DefaultLayout from './Dash/layout/DefaultLayout';
+import DashboardLayout from './Dash/layout/DashboardLayout';
 import MainLayout from './Home/layout/MainLayout';
 
 
@@ -16,7 +16,7 @@ const {signOutHandler, signedIn,setUser, userGroup, user} = props;
 
   return (
         <Switch>
-          <Route path="/dash" name="DashboardHome" render={(props) => <DefaultLayout  userGroup={userGroup} signedIn={signedIn} setUser={setUser} user={user} signOut={handleSignOut} {...props} />} />
+          <Route path="/dash" name="DashboardHome" render={(props) => <DashboardLayout  userGroup={userGroup} signedIn={signedIn} setUser={setUser} user={user} signOut={handleSignOut} {...props} />} />
           <Route path="/main" name="Main" render={(props) => <MainLayout userGroup={userGroup} signedIn={signedIn}  signOut={handleSignOut} {...props} />} />
           
           <Redirect from="/" to="/main"/>
